Add test infrastructure
Now the test are running in a systematic way by simply typing
ctest (or also make test if you are running make). For now
only the base link test has been ported, while the conservation
of kinematic energy test is broken (see issue
https://github.com/robotology-playground/mex-wholebodymodel/issues/30).
In the meanwhile also the FindMatlab.cmake was ported to the newest version
available from upstream CMake.